+++ packages/master-detail-layout/src/vaadin-master-detail-layout.js
@@ -576,78 +576,4 @@ class MasterDetailLayout extends SlotStylesMixin(ResizeMixin(ElementMixin(Themab
 
 defineCustomElement(MasterDetailLayout);
 
-class MasterDetailLayoutDetail extends ElementMixin(LitElement) {
-  static get is() {
-    return 'vaadin-master-detail-layout-detail';
-  }
-
-  static get styles() {
-    return css`
-      :host {
-        display: contents;
-      }
-    `;
-  }
-
-  constructor() {
-    super();
-    this.setAttribute('slot', 'detail-hidden');
-    this.__previousChildren = [];
-  }
-
-  /** @protected */
-  render() {
-    return html` <slot @slotchange="${this.__onSlotChange}"></slot>`;
-  }
-
-  /** @private */
-  async __onSlotChange(e) {
-    // TODO: Ignore text nodes?
-    const supportViewTransitions = typeof document.startViewTransition === 'function';
-    const nextChildren = e.target.assignedNodes();
-    const hasNextChildren = nextChildren.length > 0;
-    const hasPreviousChildren = this.__previousChildren.length > 0;
-    const layout = this.closest('vaadin-master-detail-layout');
-
-    const updateSlot = () => {
-      if (hasNextChildren) {
-        this.setAttribute('slot', 'detail');
-      } else {
-        this.setAttribute('slot', 'detail-hidden');
-      }
-
-      if (layout) {
-        layout.__updateDetailSlot();
-      }
-      this.__previousChildren = nextChildren;
-    };
-
-    if (!layout || layout.noAnimation || !supportViewTransitions || this.__transitioning) {
-      updateSlot();
-      return;
-    }
-
-    this.__transitioning = true;
-
-    // Restore previous DOM state before starting transition
-    nextChildren.forEach((child) => this.removeChild(child));
-    this.__previousChildren.forEach((child) => this.appendChild(child));
-
-    // Start view transition
-    const transitionType = hasPreviousChildren && hasNextChildren ? 'replace' : hasPreviousChildren ? 'remove' : 'add';
-    layout.setAttribute('transition', transitionType);
-    const transition = document.startViewTransition(() => {
-      // Redo DOM changes and update slot
-      nextChildren.forEach((child) => this.appendChild(child));
-      this.__previousChildren.forEach((child) => this.removeChild(child));
-      updateSlot();
-    });
-    await transition.finished;
-    layout.removeAttribute('transition');
-    this.__transitioning = false;
-  }
-}
-
-defineCustomElement(MasterDetailLayoutDetail);
-
 export { MasterDetailLayout };
