+++ packages/master-detail-layout/vaadin-master-detail-layout-detail.d.ts
@@ -1,0 +1,1 @@
+export * from './src/vaadin-master-detail-layout-detail.js';
