+++ packages/master-detail-layout/vaadin-master-detail-layout-detail.js
@@ -1,0 +1,1 @@
+export * from './src/vaadin-master-detail-layout-detail.js';
